requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==2.0.1
django-crispy-forms==1.7.0
django-filter==1.1.0
djangorestframework==3.7.7
kombu==4.1.0
Markdown==2.6.11
Pillow==5.0.0
pytz==2017.3
vine==1.1.4
